[fix] Fix the broken debug and log_named_modules flag.
Previously, because we are storing None to make it work such that
the debug flag being set to True in a config will not be overwritten,
we ignored the possibility that it would be set to False from the CLI
by not having any input.

This was fixed alongside the log_named_modules flag.