[fix] Allow querying of the text only embeddings modality.
I improved this by setting the no image prompt, and then
reusing this in different places such that we can input the
no image prompt to the database.

We also added the text prompt to not be the modified chat template,
but rather the actual inputted prompt from the config.